refactor common slotLimits functionality
