+++ webserver/server/app/controllers/DelegationForPoolController.ts
@@ -8,7 +8,11 @@ import { Routes } from '../../../shared/routes';
 import { delegationsForPool } from '../services/DelegationForPool';
 import type { DelegationForPoolResponse } from '../../../shared/models/DelegationForPool';
 import { POOL_DELEGATION_LIMIT } from '../../../shared/constants';
-import { resolvePageStart, resolveUntilTransaction } from '../services/PaginationService';
+import {
+  adjustToSlotLimits,
+  resolvePageStart,
+  resolveUntilTransaction,
+} from '../services/PaginationService';
 import { slotBoundsPagination } from '../models/pagination/slotBoundsPagination.queries';
 import { expectType } from 'tsd';
 
@@ -73,31 +77,12 @@ export class DelegationForPoolController extends Controller {
         });
       }
 
-      let pageStartWithSlot = pageStart;
-
-      // if the slotLimits field is set, this shrinks the tx id range
-      // accordingly if necessary.
-      if (requestBody.slotLimits) {
-        const bounds = slotBounds ? slotBounds[0] : { min_tx_id: -1, max_tx_id: -2 };
-
-        const minTxId = Number(bounds.min_tx_id);
-
-        if (!pageStartWithSlot) {
-          pageStartWithSlot = {
-            // block_id is not really used by this query.
-            block_id: -1,
-            // if no *after* argument is provided, this starts the pagination
-            // from the corresponding slot. This allows skipping slots you are
-            // not interested in. If there is also no slotLimits specified this
-            // starts from the first tx because of the default of -1.
-            tx_id: minTxId,
-          };
-        } else {
-          pageStartWithSlot.tx_id = Math.max(Number(bounds.min_tx_id), pageStartWithSlot.tx_id);
-        }
-
-        until.tx_id = Math.min(until.tx_id, Number(bounds.max_tx_id));
-      }
+      const pageStartWithSlot = adjustToSlotLimits(
+        pageStart,
+        until,
+        requestBody.slotLimits,
+        slotBounds
+      );
 
       const response = await delegationsForPool({
         pools: requestBody.pools.map(poolId => Buffer.from(poolId, 'hex')),
