+++ webserver/server/app/services/PaginationService.ts
@@ -2,6 +2,7 @@ import type { PoolClient } from 'pg';
 import { pageStartByHash } from '../models/pagination/pageStartByHash.queries';
 import { sqlBlockByHash } from '../models/pagination/sqlBlockByHash.queries';
 import { sqlTransactionBeforeBlock } from '../models/pagination/sqlTransactionBeforeBlock.queries';
+import type { ISlotBoundsPaginationResult } from '../models/pagination/slotBoundsPagination.queries';
 
 export type UntilPaginationType = {
   until: {
@@ -73,4 +74,43 @@ export async function resolvePageStart(request: {
     block_id: result[0].after_block_id,
     tx_id: Number.parseInt(result[0].after_tx_id, 10),
   };
+}
+
+export type SlotLimits = {
+  // this is exclusive
+  from: number;
+  // this is inclusive
+  to: number;
+};
+
+export function adjustToSlotLimits(
+  pageStartWithSlot: { tx_id: number, block_id: number } | undefined,
+  until: { tx_id: number },
+  slotLimits: SlotLimits | undefined,
+  slotBounds: ISlotBoundsPaginationResult[] | undefined,
+): { tx_id: number, block_id: number } | undefined {
+  // if the slotLimits field is set, this shrinks the tx id range
+  // accordingly if necessary.
+  if (slotLimits) {
+    const bounds = slotBounds ? slotBounds[0] : { min_tx_id: -1, max_tx_id: -2 };
+
+    const minTxId = Number(bounds.min_tx_id);
+
+    if (!pageStartWithSlot) {
+      pageStartWithSlot = {
+        block_id: -1,
+        // if no *after* argument is provided, this starts the pagination
+        // from the corresponding slot. This allows skipping slots you are
+        // not interested in. If there is also no slotLimits specified this
+        // starts from the first tx because of the default of -1.
+        tx_id: minTxId,
+      };
+    } else {
+      pageStartWithSlot.tx_id = Math.max(Number(bounds.min_tx_id), pageStartWithSlot.tx_id);
+    }
+
+    until.tx_id = Math.min(until.tx_id, Number(bounds.max_tx_id));
+  }
+
+  return pageStartWithSlot;
 }
