+++ webserver/shared/models/TransactionHistory.ts
@@ -1,3 +1,4 @@
+import { SlotLimits } from "../../server/app/services/PaginationService";
 import type { Address } from "./Address";
 import type { BlockSubset } from "./BlockLatest";
 import type { Pagination, RelationFilter } from "./common";
@@ -52,11 +53,4 @@ export type TxAndBlockInfo = {
 };
 export type TransactionHistoryResponse = {
   transactions: TxAndBlockInfo[];
-};
-
-export type SlotLimits = {
-  // this is exclusive
-  from: number;
-  // this is inclusive
-  to: number;
 };
